Updating configurations

+++ scripts/hosts-config.sh
@@ -5,8 +5,7 @@
 
 # This script configures the /etc/hosts
 # Features:
-# - Set de root password
-# - Enable sudo for professors
+# - Set initial hosts file
 
 # /soft64/admin/scripts/update-hosts.sh
 # /soft64/admin/etc/hosts_template
@@ -14,7 +13,12 @@
 HOSTNAME=$(hostname)
 IP=$(ifconfig | grep 10.32. | tr -s '[:space:]' | cut -d' ' -f3 | cut -d: -f2)
 
-# incialmente clona o arquivo pronto do soft64
-# depois o cron job vai atualizar periodicamene
-cp -f /soft64/admin/etc/hosts_current /etc/hosts
-echo "$IP   $HOSTNAME" >> /etc/hosts
+
+if [ -f /soft64/admin/etc/hosts_current ]; then
+
+	# incialmente clona o arquivo pronto do soft64
+	# depois o cron job vai atualizar periodicamene
+	cp -f /soft64/admin/etc/hosts_current /etc/hosts
+	echo "$IP   $HOSTNAME" >> /etc/hosts
+
+fi
